+++ src/ufo/GeoVaLs.interface.h
@@ -64,7 +64,7 @@ extern "C" {
   void ufo_geovals_get_f90(const F90goms &, const int &, const char *, const int &,
                            const int &, float &);
   void ufo_geovals_get_loc_f90(const F90goms &, const int &, const char *, const int &,
-                               const int &, float &);
+                               const int &, double &);
   void ufo_geovals_getdouble_f90(const F90goms &, const int &, const char *, const int &,
                            const int &, double &);
   void ufo_geovals_putdouble_f90(const F90goms &, const int &, const char *, const int &,
